:truck: Move Release Email Logic to new Directory
