+++ src/LearningObjects/Publishing/ReleaseEmails/lambda-gateway-stub.ts
